requirement.txt
accelerate==0.33.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bitsandbytes==0.43.2
cargo==0.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
distro==1.9.0
et_xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
jinja2==3.1.4
joblib==1.4.2
jq==1.7.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.1
peft==0.12.0
pillow==10.3.0
Pillow_SIMD==9.0.0.post1
psutil==5.9.8
pyarrow_hotfix==0.6
pycode-similar==1.4
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python_dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.32.3
safetensors==0.4.3
scikit_learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xxhash==3.2.0
yarl==1.9.4
